fix: Address review feedback — spinner consistency, test file convention
- draw_lead_indicator now shows spinner glyph when tool entries are
  in-progress regardless of lead_working, consistent with
  any_spinner_visible(). Added test asserting spinner appears when
  lead_working=false but entries are in-progress.
- Moved draw_lead_indicator tests from inline mod tests block to
  separate chat_tests.rs file, per CLAUDE.md convention.
